requirements.txt
boto3
streamlit
langchain
langchain-aws
langchain-community
langchain-huggingface
langchain-text-splitters
pypdf
python-dotenv
faiss-cpu
sentence-transformers
beautifulsoup4
requests

# Optional: Install based on your MODEL_PROVIDER
# For OpenAI (MODEL_PROVIDER=openai):
# openai

# For Anthropic (MODEL_PROVIDER=anthropic):
# anthropic
